+++ magnus-macros/src/typed_data.rs
@@ -1,6 +1,6 @@
 use proc_macro2::TokenStream;
 use quote::{quote, ToTokens};
-use syn::{spanned::Spanned, DeriveInput, Error, Meta};
+use syn::{spanned::Spanned, Data, DataEnum, DeriveInput, Error, Fields, Meta};
 
 use crate::util;
 
@@ -107,6 +107,71 @@ pub fn expand_derive_typed_data(input: DeriveInput) -> TokenStream {
     };
 
     let ident = input.ident;
+
+    let mut arms = Vec::new();
+    if let Data::Enum(DataEnum { variants, .. }) = input.data {
+        for variant in variants.into_iter() {
+            let mut attrs = variant
+                .attrs
+                .into_iter()
+                .filter(|attr| attr.path.is_ident("magnus"))
+                .collect::<Vec<_>>();
+            if attrs.is_empty() {
+                continue;
+            } else if attrs.len() > 1 {
+                return attrs
+                    .into_iter()
+                    .map(|a| Error::new(a.span(), "duplicate attribute"))
+                    .reduce(|mut a, b| {
+                        a.combine(b);
+                        a
+                    })
+                    .unwrap()
+                    .into_compile_error();
+            }
+            let attrs = match attrs.remove(0).parse_meta() {
+                Ok(Meta::List(v)) => v.nested.into_iter().collect(),
+                Ok(v) => return Error::new_spanned(v, "Expected meta list").into_compile_error(),
+                Err(e) => return e.into_compile_error(),
+            };
+            let mut args = match util::Args::new(attrs, &["class"]) {
+                Ok(v) => v,
+                Err(e) => return e.into_compile_error(),
+            };
+            let class = match args.extract::<String>("class") {
+                Ok(v) => v,
+                Err(e) => return e.into_compile_error(),
+            };
+            let ident = variant.ident;
+            let fetch_class = quote! {
+                *magnus::memoize!(RClass: {
+                    let class: RClass = RClass::default().funcall("const_get", (#class,)).unwrap();
+                    class.undef_alloc_func();
+                    class
+                })
+            };
+            arms.push(match variant.fields {
+                Fields::Named(_) => quote! { Self::#ident { .. } => #fetch_class },
+                Fields::Unnamed(_) => quote! { Self::#ident(_) => #fetch_class },
+                Fields::Unit => quote! { Self::#ident => #fetch_class },
+            });
+        }
+    }
+    let class_for = if !arms.is_empty() {
+        quote! {
+            fn class_for(value: &Self) -> magnus::RClass {
+                use magnus::{Module, Class, RClass};
+                #[allow(unreachable_patterns)]
+                match value {
+                    #(#arms,)*
+                    _ => Self::class(),
+                }
+            }
+        }
+    } else {
+        quote! {}
+    };
+
     let mut builder = Vec::new();
     builder.push(quote! { let mut builder = magnus::DataType::builder::<Self>(#name); });
     if mark {
@@ -145,6 +210,8 @@ pub fn expand_derive_typed_data(input: DeriveInput) -> TokenStream {
                     #builder
                 })
             }
+
+            #class_for
         }
     };
     tokens